move capsule class to db package
