+++ chainbase/src/main/java/org/tron/core/db/iterator/RockStoreIterator.java
@@ -1,4 +1,4 @@
-package org.tron.core.db.common.iterator;
+package org.tron.core.db.iterator;
 
 import java.io.IOException;
 import java.util.Map.Entry;
