+++ chainbase/src/main/java/org/tron/core/db2/common/RocksDB.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 import org.tron.common.parameter.CommonParameter;
 import org.tron.common.storage.WriteOptionsWrapper;
 import org.tron.common.storage.rocksdb.RocksDbDataSourceImpl;
-import org.tron.core.db.common.iterator.DBIterator;
+import org.tron.core.db.iterator.DBIterator;
 
 public class RocksDB implements DB<byte[], byte[]>, Flusher {
